requirements.txt
pyyaml
opencv-python
matplotlib
numpy
albumentations
tqdm
tensorboard
torchmetrics
pandas
mediapipe